requirements.txt
alembic==1.13.0
annotated-types==0.7.0
beautifulsoup4==4.14.3
black==23.12.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
dnspython==2.7.0
email-validator==2.1.0
exceptiongroup==1.3.1
extruct==0.18.0
Flask==3.0.0
Flask-Login==0.6.3
html-text==0.7.1
html5lib==1.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
jstyleson==0.0.2
loguru==0.7.2
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.10
MarkupSafe==3.0.3
mf2py==2.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.5
pyRdfa3==3.6.4
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rdflib==7.5.0
recipe_scrapers==15.10.0
requests==2.32.5
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.23
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.1
w3lib==2.3.1
webencodings==0.5.1
Werkzeug==3.1.4
zipp==3.23.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
